+++ src/sdc11073/consumer/consumerimpl.py
@@ -7,7 +7,6 @@ import dataclasses
 import functools
 import logging
 import ssl
-import traceback
 import uuid
 from typing import TYPE_CHECKING, Any
 from urllib.parse import urlparse
@@ -595,7 +594,6 @@ class SdcConsumer:
         self._subscription_mgr.start()
 
         # flag 'self.all_subscribed' tells mdib that mdib state versions shall not have any gaps
-        # => log warnings for missing versions
         self.all_subscribed = True
         not_subscribed_actions_set = set() if not_subscribed_actions is None else set(not_subscribed_actions)
         if not_subscribed_actions:
@@ -611,37 +609,6 @@ class SdcConsumer:
         for client in self._service_clients.values():
             client.set_operations_manager(self.operations_manager)
 
-        # start all subscriptions
-        # group subscriptions per hosted service
-        for dpws_hosted in self.host_description.relationship.Hosted:
-            available_actions: list[DispatchKey] = []
-            if dpws_hosted.Types is not None:
-                for port_type_qname in dpws_hosted.Types:
-                    client = self.client(port_type_qname.localname)
-                    if client is not None:
-                        available_actions.extend(client.get_available_subscriptions())
-            if len(available_actions) > 0:
-                subscribe_actions = {a for a in available_actions if a.action not in not_subscribed_actions_set}
-                if len(subscribe_actions) > 0:
-                    filter_type = eventing_types.FilterType()
-                    filter_type.text = ' '.join(x.action for x in subscribe_actions)
-                    filter_type.Dialect = DeviceEventingFilterDialectURI.ACTION
-                    try:
-                        self.do_subscribe(dpws_hosted, filter_type, subscribe_actions)
-                    except Exception:
-                        self.all_subscribed = False  # => don't log errors when mdib versions are missing
-                        self._logger.exception(  # noqa: PLE1205
-                            'start_all: could not subscribe: error = {}, actions= {}',
-                            traceback.format_exc(),
-                            subscribe_actions,
-                        )
-
-        def _update_is_connected(subscription_status: dict[str, bool]):
-            self.is_connected = all(subscription_status.values()) and any(subscription_status)
-
-        properties.strongbind(self, subscription_status=_update_is_connected)
-        _update_is_connected(self.subscription_status)
-
         # register callback for end of subscription
         self._services_dispatcher.register_post_handler(
             DispatchKey(
@@ -650,6 +617,32 @@ class SdcConsumer:
             ),
             self._on_subscription_end,
         )
+
+        # start all subscriptions - group subscriptions per hosted service
+        try:
+            for dpws_hosted in self.host_description.relationship.Hosted:
+                available_actions: list[DispatchKey] = []
+                if dpws_hosted.Types is not None:
+                    for port_type_qname in dpws_hosted.Types:
+                        client = self.client(port_type_qname.localname)
+                        if client is not None:
+                            available_actions.extend(client.get_available_subscriptions())
+                if len(available_actions) > 0:
+                    subscribe_actions = {a for a in available_actions if a.action not in not_subscribed_actions_set}
+                    if len(subscribe_actions) > 0:
+                        filter_type = eventing_types.FilterType()
+                        filter_type.text = ' '.join(x.action for x in subscribe_actions)
+                        filter_type.Dialect = DeviceEventingFilterDialectURI.ACTION
+                        self.do_subscribe(dpws_hosted, filter_type, subscribe_actions)
+        except:
+            self.stop_all()
+            raise
+
+        def _update_is_connected(subscription_status: dict[str, bool]):
+            self.is_connected = all(subscription_status.values()) and any(subscription_status)
+
+        properties.strongbind(self, subscription_status=_update_is_connected)
+        _update_is_connected(self.subscription_status)
 
     def stop_all(self, unsubscribe: bool = True):
         """Stop all threads, optionally unsubscribe."""